feat(error handling): make it easier to define custom error handlers by providing a base class with overridable virtuals

+++ Turner.Infrastructure.Crud/Requests/DeleteRequestHandler.cs
@@ -71,15 +71,6 @@ namespace Turner.Infrastructure.Crud.Requests
                 .SelectAsync(request, selector)
                 .Configure();
 
-            if (entity == null && RequestConfig.ErrorConfig.FailedToFindInDeleteIsError)
-            {
-                throw new FailedToFindException("Failed to find entity.")
-                {
-                    RequestTypeProperty = request.GetType(),
-                    QueryTypeProperty = typeof(TEntity)
-                };
-            }
-
             return entity;
         }
 
@@ -109,6 +100,10 @@ namespace Turner.Infrastructure.Crud.Requests
         public async Task<Response> HandleAsync(TRequest request)
         {
             var entity = await GetEntity(request).Configure();
+
+            if (entity == null && RequestConfig.ErrorConfig.FailedToFindInDeleteIsError)
+                return ErrorDispatcher.Dispatch(new FailedToFindError(request, typeof(TEntity)));
+
             if (entity != null)
                 await DeleteEntity(request, entity).Configure();
 
@@ -133,6 +128,12 @@ namespace Turner.Infrastructure.Crud.Requests
         {
             var entity = await GetEntity(request).Configure();
             TOut result = default(TOut);
+
+            if (entity == null && RequestConfig.ErrorConfig.FailedToFindInDeleteIsError)
+            {
+                var error = new FailedToFindError(request, typeof(TEntity), result);
+                return ErrorDispatcher.Dispatch<TOut>(error);
+            }
 
             if (entity != null)
             {
